+++ bindings/angular2/examples/tabbar.ts
@@ -1,7 +1,6 @@
 import {
   Component,
   ViewChild,
-  OnsTab,
   OnsenModule,
   NgModule,
   CUSTOM_ELEMENTS_SCHEMA