+++ bindings/react/src/index.js
@@ -31,6 +31,7 @@ import Ripple from './components/Ripple.jsx';
 import RouterNavigator from './components/RouterNavigator.jsx';
 import Row from './components/Row.jsx';
 import SearchInput from './components/SearchInput.jsx';
+import Segment from './components/Segment.jsx';
 import Select from './components/Select.jsx';
 import SpeedDial from './components/SpeedDial.jsx';
 import SpeedDialItem from './components/SpeedDialItem.jsx';
@@ -83,6 +84,7 @@ export {
   RouterUtil,
   Row,
   SearchInput,
+  Segment,
   Select,
   SpeedDial,
   SpeedDialItem,